+++ build.gradle
@@ -58,6 +58,9 @@ dependencies {
     annotationProcessor 'com.querydsl:querydsl-apt:5.0.0:jakarta'
     annotationProcessor "jakarta.annotation:jakarta.annotation-api"
     annotationProcessor "jakarta.persistence:jakarta.persistence-api"
+
+    // 7. 메일 인증 관련
+    implementation 'org.springframework.boot:spring-boot-starter-mail:3.1.5'
 }
 
 tasks.named('test') {
