+++ src/main/java/com/example/plathome/global/config/JpaConfig.java
@@ -1,6 +1,6 @@
 package com.example.plathome.global.config;
 
-import com.example.plathome.login.member.domain.UserContext;
+import com.example.plathome.login.jwt.domain.UserContext;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.domain.AuditorAware;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
