+++ src/main/java/com/example/plathome/login/email/common/EmailStaticField.java
@@ -1,0 +1,8 @@
+package com.example.plathome.login.email.common;
+
+public class EmailStaticField {
+
+    public static final long AUTH_CODE_EXPIRATION = 1000L * 60 * 30; //30min
+    public static final String TITLE = "PlatHome 이메일 인증코드 발송";
+    public static final String SEND_MESSAGE = "이메일 인증번호를 전송했습니다.";
+}
