+++ src/main/java/com/example/plathome/login/jwt/domain/SecretKey.java
@@ -1,0 +1,15 @@
+package com.example.plathome.login.jwt.domain;
+
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.stereotype.Component;
+
+import java.util.Base64;
+
+@Component
+public record SecretKey(@Value("${jwt.secretKey}") String value) {
+
+    public static final String KEY_PATH = "jwt.secretKey";
+    public byte[] getDecoded() {
+        return Base64.getDecoder().decode(value);
+    }
+}
