+++ src/main/java/com/example/plathome/login/jwt/exception/InvalidRefreshTokenException.java
@@ -1,0 +1,13 @@
+package com.example.plathome.login.jwt.exception;
+
+
+import com.example.plathome.global.exception.UnauthorizedException;
+
+import static com.example.plathome.global.error.ErrorStaticField.INVALID_REFRESH_TOKEN;
+
+public class InvalidRefreshTokenException extends UnauthorizedException {
+    private static final String MESSAGE = INVALID_REFRESH_TOKEN;
+    public InvalidRefreshTokenException() {
+        super(MESSAGE);
+    }
+}
