--- src/main/java/com/example/plathome/login/member/common/JwtStaticField.java
@@ -1,9 +1,0 @@
-package com.example.plathome.login.member.common;
-
-public class JwtStaticField {
-    public static final String BEARER = "Bearer ";
-    public static final String REFRESH_URL = "/token";
-
-    public static final long ACCESS_TOKEN_EXPIRATION = 1000L * 60 * 30; // Access token is 15 minutes.
-    public static final long REFRESH_TOKEN_EXPIRATION = 1000L * 60 * 60 * 24; // Refresh token is one day.
-}
